Fixed multiple errors where ai-toolkit didn't run on macOS.
- Added several missing tests for MPS which caused CUDA errors.

+++ extensions_built_in/diffusion_models/chroma/src/math.py
@@ -33,7 +33,8 @@ def attention(q: Tensor, k: Tensor, v: Tensor, pe: Tensor, mask: Tensor) -> Tens
 
 def rope(pos: Tensor, dim: int, theta: int) -> Tensor:
     assert dim % 2 == 0
-    scale = torch.arange(0, dim, 2, dtype=torch.float64, device=pos.device) / dim
+    # Use float32 for MPS compatibility (MPS doesn't support float64)
+    scale = torch.arange(0, dim, 2, dtype=torch.float32, device=pos.device) / dim
     omega = 1.0 / (theta**scale)
     out = torch.einsum("...n,d->...nd", pos, omega)
     out = torch.stack(
